- Việt hóa trang web - Tự động thêm số lượng đã bán của sản phẩm và giảm tồn kho của sản phẩm - Nếu sản phẩm hết hàng thì sẽ ẩn nút thêm vào giỏ hàng ngoài home và ẩn nút thêm giỏ hàng trong singleproduct - xây dựng giao diện quản lí

+++ product/models.py
@@ -12,7 +12,7 @@ class Product(models.Model):
     title = models.CharField(default='',max_length=255)
     price  = models.IntegerField(default=0)
     description = models.TextField(default='', blank=True)
-    amount = models.FloatField(default=0)
+    amount = models.IntegerField(default=0)
     active = models.BooleanField(default=True)
     date_add = models.DateField(auto_now_add=True)
     category = models.ForeignKey(Category, on_delete=models.CASCADE)
